CPipewireRegistry: move globals from CPipewire
