+++ xbmc/cores/AudioEngine/Sinks/pipewire/Pipewire.h
@@ -8,13 +8,7 @@
 
 #pragma once
 
-#include "cores/AudioEngine/Sinks/pipewire/PipewireProxy.h"
-
-#include <map>
 #include <memory>
-#include <string>
-
-#include <pipewire/properties.h>
 
 namespace AE
 {
@@ -43,28 +37,7 @@ public:
   CPipewireRegistry* GetRegistry() { return m_registry.get(); }
   std::shared_ptr<CPipewireStream>& GetStream() { return m_stream; }
 
-  struct PipewirePropertiesDeleter
-  {
-    void operator()(pw_properties* p) { pw_properties_free(p); }
-  };
-
-  struct global
-  {
-    std::string name;
-    std::string description;
-    uint32_t id;
-    uint32_t permissions;
-    std::string type;
-    uint32_t version;
-    std::unique_ptr<pw_properties, PipewirePropertiesDeleter> properties;
-    std::unique_ptr<CPipewireProxy> proxy;
-  };
-
-  std::map<uint32_t, std::unique_ptr<global>>& GetGlobals() { return m_globals; }
-
 private:
-  std::map<uint32_t, std::unique_ptr<global>> m_globals;
-
   std::unique_ptr<CPipewireThreadLoop> m_loop;
   std::unique_ptr<CPipewireContext> m_context;
   std::unique_ptr<CPipewireCore> m_core;
