+++ xbmc/cores/AudioEngine/Sinks/pipewire/PipewireRegistry.h
@@ -8,9 +8,14 @@
 
 #pragma once
 
+#include "cores/AudioEngine/Sinks/pipewire/PipewireProxy.h"
+
+#include <map>
 #include <memory>
+#include <string>
 
 #include <pipewire/core.h>
+#include <pipewire/properties.h>
 
 namespace AE
 {
@@ -29,6 +34,24 @@ public:
   pw_registry* Get() const { return m_registry.get(); }
 
   void AddListener(void* userdata);
+  struct PipewirePropertiesDeleter
+  {
+    void operator()(pw_properties* p) { pw_properties_free(p); }
+  };
+
+  struct global
+  {
+    std::string name;
+    std::string description;
+    uint32_t id;
+    uint32_t permissions;
+    std::string type;
+    uint32_t version;
+    std::unique_ptr<pw_properties, PipewirePropertiesDeleter> properties;
+    std::unique_ptr<CPipewireProxy> proxy;
+  };
+
+  std::map<uint32_t, std::unique_ptr<global>>& GetGlobals() { return m_globals; }
 
 private:
   static void OnGlobalAdded(void* userdata,
@@ -50,6 +73,8 @@ private:
   };
 
   std::unique_ptr<pw_registry, PipewireRegistryDeleter> m_registry;
+
+  std::map<uint32_t, std::unique_ptr<global>> m_globals;
 };
 
 } // namespace PIPEWIRE
